pyproject.toml
[build-system]
build-backend = "hatchling.build"
requires = [
    "hatchling",
]

[project]
name = "aa-sov-timer"
version = "1.12.2"
description = "Sov Campaign Timer for Alliance Auth"
readme = "README.md"
keywords = [
    "allianceauth",
    "eveonline",
    "sovereignty_timer",
    "timerboard",
]
license = {file = "LICENSE"}
authors = [
    { name = "Peter Pfeufer", email = "develop@ppfeufer.de" },
]
requires-python = "~=3.8"
classifiers = [
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Django :: 4.0",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
]
dependencies = [
    "allianceauth>=3.2",
    "allianceauth-app-utils>=1.12",
    "django-eveuniverse>=0.14",
]
[project.urls]
Changelog = "https://github.com/ppfeufer/aa-sov-timer/blob/master/CHANGELOG.md"
Documentation = "https://github.com/ppfeufer/aa-sov-timer/blob/master/README.md"
Donations = "https://ko-fi.com/ppfeufer"
Homepage = "https://github.com/ppfeufer/aa-sov-timer"
Source = "https://github.com/ppfeufer/aa-sov-timer.git"
Tracker = "https://github.com/ppfeufer/aa-sov-timer/issues"

[tool.hatch.build]
include = [
    "/sovtimer",
]
